fix: Maintain img size regardless of its state via grid + svg was not working bc of lowercase b in viewbox
This was in part due to HTML linter. It only has error level severity, eslint style warn severity support is in the works,
see: https://github.com/htmlhint/HTMLHint/issues/279#issuecomment-417939741

The other part was XML vs HTML fragment, latter's attributes are always set as lowercase! Which broke the SVG! Reference for fix is found in inline comment in code.